refactor: adopt upstream secrets collision fix
